sqlserver/requirements.txt
pyodbc>=4.0.39
mcp[cli]>=1.5.0
python-dotenv>=1.0.0